# src/main/resources/api-docs.json
{"openapi":"3.0.1","info":{"title":"Ordo","version":"0.0.1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/workspaces/{id}":{"get":{"tags":["workspace-controller"],"operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WorkspaceRead"}}}}}},"put":{"tags":["workspace-controller"],"operationId":"updateWorkspace","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"dto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/WorkspaceUpdate"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WorkspaceRead"}}}}}},"delete":{"tags":["workspace-controller"],"operationId":"deleteWorkspace","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WorkspaceRead"}}}}}}},"/tasks/{id}":{"get":{"tags":["task-controller"],"operationId":"findById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaskRead"}}}}}},"put":{"tags":["task-controller"],"operationId":"updateTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaskRead"}}}}}},"delete":{"tags":["task-controller"],"operationId":"deleteTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaskRead"}}}}}}},"/lists/{id}":{"get":{"tags":["list-controller"],"operationId":"findById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ListRead"}}}}}},"put":{"tags":["list-controller"],"operationId":"updateTaskList","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListCreate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ListRead"}}}}}},"delete":{"tags":["list-controller"],"operationId":"deleteTaskList","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ListRead"}}}}}}},"/comments/{id}":{"get":{"tags":["comment-controller"],"operationId":"findById_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentRead"}}}}}},"put":{"tags":["comment-controller"],"operationId":"updateComment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentCreate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentRead"}}}}}},"delete":{"tags":["comment-controller"],"operationId":"deleteComment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentRead"}}}}}}},"/cards/{id}":{"get":{"tags":["card-controller"],"operationId":"findById_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardWithItsListRead"}}}}}},"put":{"tags":["card-controller"],"operationId":"updateTask_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardCreate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardWithItsListRead"}}}}}},"delete":{"tags":["card-controller"],"operationId":"deleteTask_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardWithItsListRead"}}}}}}},"/boards/{id}":{"get":{"tags":["board-controller"],"operationId":"findById_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BoardRead"}}}}}},"put":{"tags":["board-controller"],"operationId":"updateBoard","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"dto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/BoardCreate"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BoardRead"}}}}}},"delete":{"tags":["board-controller"],"operationId":"deleteBoard","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BoardRead"}}}}}}},"/boards/{id}/members/{newUserId}/{roleId}":{"put":{"tags":["board-controller"],"operationId":"addMember","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"newUserId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"roleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BoardRead"}}}}}}},"/boards/{id}/members/{memberId}/{newRoleId}":{"put":{"tags":["board-controller"],"operationId":"updateMemberRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"memberId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"newRoleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BoardRead"}}}}}}},"/workspaces":{"get":{"tags":["workspace-controller"],"operationId":"findAllByOwner","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageWorkspaceRead"}}}}}},"post":{"tags":["workspace-controller"],"operationId":"createWorkspace","parameters":[{"name":"dto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/WorkspaceCreate"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WorkspaceRead"}}}}}}},"/tasks":{"post":{"tags":["task-controller"],"operationId":"createTask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaskRead"}}}}}}},"/lists":{"post":{"tags":["list-controller"],"operationId":"createTaskList","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListCreate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ListRead"}}}}}}},"/comments":{"post":{"tags":["comment-controller"],"operationId":"createComment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentCreate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentRead"}}}}}}},"/cards":{"post":{"tags":["card-controller"],"operationId":"createTask_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardCreate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardWithItsListRead"}}}}}}},"/boards":{"get":{"tags":["board-controller"],"operationId":"findAllAccessibleBoards","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageBoardRead"}}}}}},"post":{"tags":["board-controller"],"operationId":"createBoard","parameters":[{"name":"dto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/BoardCreate"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BoardRead"}}}}}}},"/auth/oauth2/google":{"post":{"tags":["o-auth-2-controller"],"operationId":"exchangeGoogleCode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OAuth2CodeOnTokenRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/auth/local/register":{"post":{"tags":["local-auth-controller"],"operationId":"register","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/UserCreateDTO"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserReadDTO"}}}}}}},"/auth/local/login":{"post":{"tags":["local-auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocalLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/workspaces/{id}/roles":{"get":{"tags":["workspace-controller"],"operationId":"findRolesByWorkspaceId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageWorkspaceRoleRead"}}}}}}},"/workspaces/{id}/boards":{"get":{"tags":["workspace-controller"],"operationId":"findBoardsByWorkspaceId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageBoardRead"}}}}}}},"/secured":{"get":{"tags":["auth-controller"],"operationId":"securedEndpoint","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/lists/{id}/cards":{"get":{"tags":["list-controller"],"operationId":"findAllByListPaged","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCardRead"}}}}}}},"/comments/user":{"get":{"tags":["comment-controller"],"operationId":"findAllByUser","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCommentRead"}}}}}}},"/cards/{id}/tasks":{"get":{"tags":["card-controller"],"operationId":"findAllByCard","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageTaskRead"}}}}}}},"/cards/{id}/comments":{"get":{"tags":["card-controller"],"operationId":"findAllCommentsByCard","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCommentRead"}}}}}}},"/cards/assigned":{"get":{"tags":["card-controller"],"operationId":"findAllAssignedToUser","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCardWithItsListRead"}}}}}}},"/boards/{id}/lists":{"get":{"tags":["board-controller"],"operationId":"findAllTaskListsByBoardId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ListRead"}}}}}}}},"/boards/{id}/lists/page":{"get":{"tags":["board-controller"],"operationId":"findAllTaskListsByBoardId_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageListRead"}}}}}}},"/":{"get":{"tags":["auth-controller"],"operationId":"endpoint","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"WorkspaceUpdate":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}}},"UserReadDTO":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"},"imageUrl":{"type":"string"}}},"WorkspaceRead":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"owner":{"$ref":"#/components/schemas/UserReadDTO"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/WorkspaceRoleRead"}},"createdAt":{"type":"string","format":"date-time"}}},"WorkspaceRoleRead":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"ableToManageMembers":{"type":"boolean"},"ableToManageContent":{"type":"boolean"},"ableToManageSettings":{"type":"boolean"},"ableToManageRoles":{"type":"boolean"}}},"TaskCreate":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"position":{"type":"integer","format":"int32"},"completed":{"type":"boolean"},"cardId":{"type":"integer","format":"int64"}}},"TaskRead":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"completed":{"type":"boolean"},"position":{"type":"integer","format":"int32"}}},"ListCreate":{"type":"object","properties":{"title":{"type":"string"},"position":{"type":"integer","format":"int32"},"color":{"type":"string"},"boardId":{"type":"integer","format":"int64"}}},"BoardRead":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"visibility":{"type":"string","enum":["PUBLIC","PRIVATE","WORKSPACE"]},"createdAt":{"type":"string","format":"date-time"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/BoardRoleRead"}},"workspace":{"$ref":"#/components/schemas/WorkspaceRead"}}},"BoardRoleRead":{"type":"object","properties":{"ableToEdit":{"type":"boolean"},"ableToDelete":{"type":"boolean"},"ableToInviteMembers":{"type":"boolean"},"ableToManageRoles":{"type":"boolean"},"ableToCreateLists":{"type":"boolean"},"ableToCreateTasks":{"type":"boolean"}}},"ListRead":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"position":{"type":"integer","format":"int32"},"color":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"board":{"$ref":"#/components/schemas/BoardRead"}}},"CommentCreate":{"type":"object","properties":{"message":{"type":"string"},"cardId":{"type":"integer","format":"int64"}}},"CardWithItsListRead":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"dueDate":{"type":"string","format":"date"},"position":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"taskList":{"$ref":"#/components/schemas/ListRead"},"createdBy":{"$ref":"#/components/schemas/UserReadDTO"},"assignedTo":{"$ref":"#/components/schemas/UserReadDTO"}}},"CommentRead":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"message":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"createdBy":{"$ref":"#/components/schemas/UserReadDTO"},"card":{"$ref":"#/components/schemas/CardWithItsListRead"}}},"CardCreate":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"dueDate":{"type":"string","format":"date"},"position":{"type":"integer","format":"int32"},"listId":{"type":"integer","format":"int64"},"assignedToId":{"type":"integer","format":"int64"}}},"BoardCreate":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"visibility":{"type":"string","enum":["PUBLIC","PRIVATE","WORKSPACE"]},"workspaceId":{"type":"integer","format":"int64"}}},"WorkspaceCreate":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}}},"OAuth2CodeOnTokenRequest":{"type":"object","properties":{"code":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"user":{"$ref":"#/components/schemas/UserReadDTO"}}},"UserCreateDTO":{"type":"object","properties":{"email":{"type":"string"},"name":{"type":"string"},"password":{"type":"string"}}},"LocalLoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32","default":0},"size":{"minimum":1,"type":"integer","format":"int32","default":10}}},"PageWorkspaceRead":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/WorkspaceRead"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"unpaged":{"type":"boolean"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}},"PageWorkspaceRoleRead":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/WorkspaceRoleRead"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageBoardRead":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/BoardRead"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"CardRead":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"dueDate":{"type":"string","format":"date"},"position":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"createdBy":{"$ref":"#/components/schemas/UserReadDTO"},"assignedTo":{"$ref":"#/components/schemas/UserReadDTO"}}},"PageCardRead":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/CardRead"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageCommentRead":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/CommentRead"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageTaskRead":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/TaskRead"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageCardWithItsListRead":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/CardWithItsListRead"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageListRead":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ListRead"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}